# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BudgetWise API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense, transfer]
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
                  total:
                    type: integer
                required: [data, total]
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionBody"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        categoryId:
          type: integer
        amount:
          type: number
        type:
          type: string
          enum: [income, expense, transfer]
        description:
          type: string
        date:
          type: string
          format: date
        isRecurring:
          type: boolean
          default: false
        accountName:
          type: string
        categoryName:
          type: string
        categoryColor:
          type: string
        categoryIcon:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, accountId, amount, type, date, createdAt]
    CreateTransactionBody:
      type: object
      properties:
        accountId:
          type: integer
        categoryId:
          type: integer
        amount:
          type: number
        type:
          type: string
          enum: [income, expense, transfer]
        description:
          type: string
        date:
          type: string
          format: date
        isRecurring:
          type: boolean
          default: false
      required: [accountId, amount, type, date]
    UpdateTransactionBody:
      type: object
      properties:
        accountId:
          type: integer
        categoryId:
          type: integer
        amount:
          type: number
        type:
          type: string
          enum: [income, expense, transfer]
        description:
          type: string
        date:
          type: string
          format: date
        isRecurring:
          type: boolean

    # Budget
...